refactor: 상수명 변경
SEPARATOR -> DELIMITER
