+++ src/main/java/racingcar/constant/Delimiter.java
@@ -1,0 +1,9 @@
+package racingcar.constant;
+
+public interface Delimiter {
+
+    char DELIMITER = ',';
+    String REGEX_DELIMITER= ",";
+    String CONSECUTIVE_DELIMITER = ",,";
+    String KOREAN_BRACKET_DELIMITER = "콤마(',')";
+}
