--- src/main/java/racingcar/constant/Separator.java
@@ -1,9 +1,0 @@
-package racingcar.constant;
-
-public interface Separator {
-
-    char SEPARATOR = ',';
-    String REGEX_SEPARATOR = ",";
-    String CONSECUTIVE_SEPARATOR = ",,";
-    String KOREAN_BRACKET_SEPARATOR = "콤마(',')";
-}
