+++ src/main/java/racingcar/domain/racing/Judgement.java
@@ -5,6 +5,8 @@ import racingcar.domain.car.Car;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static racingcar.constant.Delimiter.DELIMITER;
+
 public class Judgement {
 
     public String determineWinners(final Racing racing) {
@@ -14,7 +16,7 @@ public class Judgement {
         return cars.stream()
                 .filter(car -> car.getForwardCount() == maxForwardCount)
                 .map(Car::getName)
-                .collect(Collectors.joining(", "));
+                .collect(Collectors.joining(DELIMITER + " "));
     }
 
     private int calculateMaxForwardCount(final List<Car> cars) {
